Make paths relative to scad source files
OpenSCAD looks for <includes> and <use> relative to the directory of the source
file and the OpenSCAD binary. The sheet and stl generation depended on
the second location, and therefore broke when the OpenSCAD binary was
not in the Mendel90 folder. This is the case for example if OpenSCAD is
installed in Ubuntu from the official PPA.

See also:
http://rocklinux.net/pipermail/openscad/2012-January/002235.html